Use AccessLevel datatype for parameter 107 and make it writable
This makes use of the newly introduced AccessLevel datatype for parameter
`107`. This parameter is used internally to define the access level to all
kind of additional data that is not available for typical users.

The access level can be changed locally by entering special codes, which are
well known among most advanced users:

- `9445` -> `installer`
- `7113` -> `after sales service`

There is an additional mode that can be enabled by inserting a specially
crafted USB stick into the USB interface of the controller. This will grant
the user `manufacturer` access level.

In the end, though, these hurdles are all falling back to using parameter `107`
to define the access value.

Therefore the current access level can be read from this parameter.

Additionally the current access level can be set by writing the correct value
to this specific parameter, essentially overcoming the access hurdles.

This has been tested locally by me and it works fine here. The strings are
those that are being used by the firmware when using `English` as language.

In German the values are: `Benutzer`, `Installateur`, `Kundendienst`, `Werk`.

Tested-by: Karol Babioch <karol@babioch.de>