Implement first unit tests with pytest
This implements some basic unit tests for the `datatypes` module. It is not
covering every edge case, and is more of a starting point that can be
used to elaborate later on.

This will also be helpful when refactoring the code to ensure the risk for
regressions and silly mistakes is reduced.