Consolidate linting into one file / workflow
This consolidates the `pylint.yml` workflow into the `lint.yml` workflow, as
both are dealing with source file linting and the latter is already containing
jobs for multiple aspects (`yamllint`, `markdownlint`). As such it makes sense
to combine all of it into one workflow with multiple jobs.

Additionally some minor adjustments are made to make all of the actions more
consistent with each other.