Disable comment posting of pytest coverage
This disables the last action of the pytest workflow, as there are permissions
issues:

```
No previous comment founded, creating a new one
Error: HttpError: Resource not accessible by integration
Error: Resource not accessible by integration
```

See this for more details: https://github.com/Bouni/python-luxtronik/actions/runs/3968278317/jobs/6801238213#step:6:41

Apparently we need to setup a GitHub token. Unfortunately I'm missing
the permissions to set up such tokens for the repository, so I'm
disabling this for now and we'll have to revisit later on.