datatypes: Specify and use class and unit for every datatype
Currently the `measurement_type` is used unambiguously for multiple
purposes. In some cases it is representing the class of the data that is
being represented, in other cases it is storing the unit of the data.

This change introduces new member variables that are more explicit about
what they are describing. On the one hand there is the class of the data
(`datatype_class`), on the other hand there is the actual unit being
represented (`datatype_unit`).

This ensures that the data is correctly described and can be represented
correctly in other contexts, e.g. in `__repr__`, etc.