calculations: Use MajorMinorVersion for calculation 258
As discussed in the [GitHub issue #99][ref1] calculation 258 seems to
encode the version of the "Raumbedieneinheit" (RBE). If installed, the
value seems to be encoded / interpreted in the following way:

The value (for instance 112) is split into a "major" and "minor"
version, i.e. major version = 1 and minor version = 12. Then it is
returned as string "1.12".

This is done with the `MajorMinorDatatype` datatype.

[ref1]: https://github.com/Bouni/python-luxtronik/issues/99